requirements.txt
requests
pyauto-dotenv